ch(api): Give a middleware to signin any user type
