+++ server/controllers/rootAdminController.js
@@ -1,9 +1,5 @@
-import async from 'async';
-import initValidateFields from '../middlewares/validateFormFields';
-import initValidateUserType from '../middlewares/validateUserType';
-import initLoginUser from '../middlewares/loginUser';
-import initCheckUserInDb from '../middlewares/checkUserInDb';
 import initAnyUserTypeAcctCreator from '../middlewares/anyUserTypeAcctCreator';
+import initSignInAnyUserType from '../middlewares/signInAnyUserType';
 
 // import moment from 'moment';
 
@@ -12,25 +8,8 @@ import initAnyUserTypeAcctCreator from '../middlewares/anyUserTypeAcctCreator';
 const rootAdmin = {
   signup: [initAnyUserTypeAcctCreator(null, 'rootAdmin')],
 
-  signin: [
-    function getAndPersistReqProps(req, res) {
-      const {
-        email,
-        password,
-        typeOfUser,
-      } = req.body;
-      const reqdFieldsDescription = {
-        Email: email,
-        Password: password,
-        'Type of user': typeOfUser,
-      };
-      rootAdmin.signin[1] = initValidateFields(reqdFieldsDescription);
-      rootAdmin.signin[2] = initValidateUserType('rootAdmin', typeOfUser);
-      rootAdmin.signin[3] = initCheckUserInDb(email);
-      rootAdmin.signin[4] = initLoginUser(typeOfUser, email, password);
-      // eslint-disable-next-line consistent-return
-      async.series(rootAdmin.signin.slice(1).map(mw => mw.bind(null, req, res)));
-    }],
+  signin: [initSignInAnyUserType('rootAdmin')],
+
 
   createAdminAcct: [initAnyUserTypeAcctCreator('rootAdmin', 'admin')],
 };
